Add Api Versioning, New Endpoints, Error Handling
* Add change name detector module (#63)

* Add ChangeNameDetector Module With Message Brocker

* Update Readme

* fix memory leak in character analyser


* Add Api Versioning and Filtered Characters Endpoint


* Add Error Handling, Prompt Endpoint


* Change CorrelationId from int to long


* Add comments and validations to endpoints


* Set Environment Variable for Unit Tests


* Revert "Set Environment Variable for Unit Tests"

This reverts commit 819f4391f0d626b18d15b91f9fd4a15556f1147f.

* Delete Innecesery WorldScans


---------

+++ API/src/TibiaEnemyOtherCharactersFinder.Domain/Entities/CharacterCorrelation.cs
@@ -5,7 +5,7 @@ public class CharacterCorrelation : IEntity
     /// <summary>
     /// Id of specific correlation between two characters
     /// </summary>
-    public int CorrelationId { get; set; }
+    public long CorrelationId { get; set; }
 
     /// <summary>
     /// Id of specific character that logout
